refactor: queryRunner init method refactoring
utils 폴더 하위에 공통 메서드 분리
