+++ src/utils/index.ts
@@ -1,0 +1,2 @@
+export { default as getOrCreateCategory } from './getOrCreateCategory';
+export { default as init } from './init';
